page size uint; s3compat typo fix

+++ ais/tests/common.go
@@ -48,7 +48,7 @@ var (
 	numfiles               int
 	numworkers             int
 	match                  = ".*"
-	pagesize               int64
+	pagesize               uint
 	fnlen                  int
 	prefix                 string
 	abortonerr             = false
